third_party/WebKit/LayoutTests/fast/dom/Window/dispatchEvent-expected.txt

Test window.dispatchEvent().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.dispatchEvent(event) threw exception InvalidStateError: Failed to execute 'dispatchEvent' on 'EventTarget': The event provided is null..
PASS window.dispatchEvent(event) threw exception InvalidStateError: Failed to execute 'dispatchEvent' on 'EventTarget': The event provided is uninitialized..
PASS myEventDispatched is true
PASS target is window
PASS currentTarget is window
PASS useCaptureDispatched is true
PASS nonUseCaptureDispatched is true
PASS successfullyParsed is true

TEST COMPLETE
